+++ src/Discussion/DiscussionService.php
@@ -154,17 +154,22 @@ class DiscussionService extends AbstractService
      * Get the discussion's messages list
      *
      * @param int $discussionId
+     * @param bool $markMessagesAsRead
      *
      * @return Message[]
      * @throws AuthenticationRequired
      * @throws \GuzzleHttp\Exception\GuzzleException
      * @throws \Wizaplace\SDK\Exception\JsonDecodingError
      */
-    public function getMessages(int $discussionId): array
+    public function getMessages(int $discussionId, ?bool $markMessagesAsRead = null): array
     {
         $this->client->mustBeAuthenticated();
-
-        $messages = $this->client->get('discussions/' . $discussionId . '/messages');
+        $endpoint = 'discussions/' . $discussionId . '/messages';
+        if(\is_bool($markMessagesAsRead) === true) {
+            $markMessagesAsRead = $markMessagesAsRead === true ? 'true' : 'false';
+            $endpoint .= '/?markMessagesAsRead=' . $markMessagesAsRead;
+        }
+        $messages = $this->client->get($endpoint);
         $userId = $this->client->getApiKey()->getId();
 
         return array_map(